+++ GUI+Client/GUI+Client/ShowPointsWindow.h
@@ -18,8 +18,8 @@ private:
 	Ui::ShowPointsWindowClass ui;
 	void SortPlayers();
 	void DisplayPlayers();
-	std::vector<std::pair<std::string, uint16_t>> m_players;
-	std::vector<float> m_playersPoints;
+	std::vector<std::pair<std::string,int >> m_players;
+	std::vector<int> m_playersPoints;
 	std::vector<std::string> m_playersNames;
 	Client m_client;
 };
